+++ paws.common/R/tags.R
@@ -34,6 +34,7 @@
 #' @name tags
 NULL
 
+# Returns the given tag on an object, or "" if not present.
 #' @rdname tags
 #' @export
 tag_get <- function(object, tag) {
@@ -42,6 +43,14 @@ tag_get <- function(object, tag) {
     return(tags[[tag]])
   }
   return("")
+}
+
+# Returns all tags on an object as a list.
+#' @rdname tags
+#' @export
+tag_get_all <- function(object) {
+  tags <- attr(object, "tags", exact = TRUE)
+  return(tags)
 }
 
 # Returns whether the object has the given tag.
@@ -110,7 +119,22 @@ type <- function(object) {
       "scalar"
     )
   } else {
-    t <- ""
+    t <- guess_type(object)
+  }
+  return(t)
+}
+
+# If an object has no tag information, guess the type of the object based
+# on its properties.
+guess_type <- function(object) {
+  if (!is.null(names(object))) {
+    t <- "structure"
+  } else {
+    if (is.atomic(object) && length(object) == 1) {
+      t <- "scalar"
+    } else {
+      t <- "list"
+    }
   }
   return(t)
 }