+++ paws.common/tests/testthat/test_struct.R
@@ -1,3 +1,5 @@
+context("Structs")
+
 test_that("struct constructors", {
   Foo <- struct(a = "", b = NULL)
   bar <- Foo()