+++ paws.common/tests/testthat/test_tags.R
@@ -2,10 +2,12 @@ context("Tags")
 
 test_that("add tags to an object without tags", {
   a <- list()
-  b <- tag_add(a, list(foo = "abc", bar = 123, baz = list(qux = "xyz")))
+  tags <- list(foo = "abc", bar = 123, baz = list(qux = "xyz"))
+  b <- tag_add(a, tags)
   expect_equal(tag_get(b, "foo"), "abc")
   expect_equal(tag_get(b, "bar"), 123)
   expect_equal(tag_get(b, "baz"), list(qux = "xyz"))
+  expect_equal(tag_get_all(b), tags)
 })
 
 test_that("add tags to an object with tags", {
@@ -73,4 +75,19 @@ test_that("delete some tags", {
   expect_equal(tag_get(result$inner[[1]], "foo2"), "")
   expect_equal(tag_get(result$inner[[2]], "foo2"), "")
   expect_equal(tag_get(result$inner[[3]], "foo2"), "")
+})
+
+#-------------------------------------------------------------------------------
+
+test_that("types", {
+  expect_equal(type(List()), "list")
+  expect_equal(type(Map()), "map")
+  expect_equal(type(Scalar()), "scalar")
+  expect_equal(type(Structure(Foo = logical(0))), "structure")
+
+  expect_equal(type(c(1, 2)), "list")
+  expect_equal(type(list(1, 2, 3)), "list")
+  expect_equal(type(1), "scalar")
+  expect_equal(type(c(a = 1, b = 2, c = 3)), "structure")
+  expect_equal(type(list(a = 1, b = 2, c = 3)), "structure")
 })